requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

streamlit~=1.35.0
setuptools~=69.5.1
paperswithcode-client~=0.3.1